fix: restore DelayClient interface
